Integrate support for dispatch service
The PROV-JSON loader collects a set of a events during the evaluation
and returned with the statements themselves. The resource.ImportResource
resource takes those events and sends a POST request with the
"resource-merge" event payload to the dispatch service.

The dispatch service was introduced to send notifications and triggers
service hooks. Read more about the dispatch service:
https://github.com/chop-dbhi/origins-dispatch/

Note, this integration is pending a refactor while the statement vs. command
distinction is being reconciled.

Signed-off-by: Byron Ruth <b@devel.io>